pyproject.toml
[tool.poetry]
name = "ioet-houses-bot"
version = "0.1.0"
description = ""
authors = ["Gustavo Eguez <eguezgustavo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.21.2"
pillow = "^11.0.0"
pytesseract = "^0.3.13"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
